Move listRecordFiles out of conversion into a dedicated module
listRecordFiles is a generic record-file listing helper, not specific to
file format conversion, so extract it to src/background/file/list.ts.
