+++ src/tests/background/file/conversion.spec.ts
@@ -1,4 +1,4 @@
-import { convertRecordFiles, listRecordFiles } from "@/background/file/conversion.js";
+import { convertRecordFiles } from "@/background/file/conversion.js";
 import { RecordFileFormat } from "@/common/file/record.js";
 import {
   DestinationType,
@@ -14,36 +14,6 @@ import { saveAppSettings } from "@/background/settings.js";
 import { getTempPathForTesting } from "@/background/proc/env.js";
 
 const tmpdir = path.join(getTempPathForTesting(), "conversion");
-
-describe("listRecordFiles", () => {
-  it("with-subdirectories", async () => {
-    const files = await listRecordFiles(
-      "src/tests/testdata/conversion/input",
-      [RecordFileFormat.KIF, RecordFileFormat.KIFU],
-      true,
-    );
-    expect(
-      files.map((file) => path.relative("src/tests/testdata/conversion/input", file)).sort(),
-    ).toStrictEqual([
-      "kif-sjis.kif",
-      "kifu-utf8.kifu",
-      path.join("sub01", "kif-sjis.kif"),
-      path.join("sub01", "sub0101", "kif-sjis.kif"),
-      path.join("sub02", "kif-sjis.kif"),
-    ]);
-  });
-
-  it("without-subdirectories", async () => {
-    const files = await listRecordFiles(
-      "src/tests/testdata/conversion/input",
-      [RecordFileFormat.CSA],
-      false,
-    );
-    expect(
-      files.map((file) => path.relative("src/tests/testdata/conversion/input", file)),
-    ).toStrictEqual(["csa-sjis.csa"]);
-  });
-});
 
 describe("conversion", () => {
   it("individual", async () => {
